Layer background and sidebar

+++ bg_gen/src/config.rs
@@ -3,8 +3,8 @@ use std::process;
 use crate::pause::pause;
 
 pub struct Config {
-  pub width: u16,
-  pub height: u16,
+  pub width: u32,
+  pub height: u32,
 }
 
 impl Config {
@@ -13,12 +13,12 @@ impl Config {
       return Err("Not enough arguments");
     }
 
-    let width: u16 = u16::from_str_radix(&args[1], 10).unwrap_or_else(|_| {
+    let width: u32 = u32::from_str_radix(&args[1], 10).unwrap_or_else(|_| {
       eprintln!("Invalid width");
       pause();
       process::exit(1);
     });
-    let height: u16 = u16::from_str_radix(&args[2], 10).unwrap_or_else(|_| {
+    let height: u32 = u32::from_str_radix(&args[2], 10).unwrap_or_else(|_| {
       eprintln!("Invalid height");
       pause();
       process::exit(1);
